Add support for koji profiles.
Koji profiles can be installed in system locations or a user config
directly, and optionally, the user can supply an additional
file/directory to override default settings that may be in the system
locations.  This patch enables both of these to be passed in, in
addition to the simple case of having a url to perform read-only
operations against a koji instance.

Fixes #37

Signed-off-by: Jason Guiditta <jason.guiditta@gmail.com>